+++ lib/main.dart
@@ -5,9 +5,9 @@ import 'package:todo_list/config/theme/theme.dart';
 import 'package:todo_list/core/helpers/formatting_helper.dart';
 import 'package:todo_list/features/todo/domain/bloc/todo_list_bloc.dart';
 import 'package:todo_list/features/todo/domain/bloc/todo_list_event.dart';
-import 'package:todo_list/features/todo/presentation/todo_all/screens/home_screen.dart';
-import 'package:todo_list/features/todo/presentation/todo_single/screens/single_todo_screen.dart';
-import 'package:todo_list/features/todo/presentation/todo_single/utility/todo_action.dart';
+import 'package:todo_list/features/todo/presentation/todo_all/screen/home_screen.dart';
+import 'package:todo_list/features/todo/presentation/todo_single/screen/single_todo_screen.dart';
+import 'package:todo_list/features/todo/presentation/common/todo_action.dart';
 import 'package:todo_list/features/todo/data/isar_todo_repository.dart';
 
 import 'package:todo_list/core/services/persistence_service.dart';
